data-lake-pipeline/requirements.txt
boto3
pyspark
apache-airflow